pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vigilante-geotecnico"
version = "1.0.0"
description = "Sistema de monitoreo inteligente para deformaciones geotécnicas con IA"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Tu Nombre", email = "tu.email@ejemplo.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
requires-python = ">=3.8"
dependencies = [
    "pandas",
    "numpy",
    "matplotlib",
    "requests",
    "jupyter",
    "python-dotenv",
    "openai",
    "rich",
    "fastapi",
    "uvicorn",
    "agno",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/tu-usuario/vigilante-geotecnico"
Repository = "https://github.com/tu-usuario/vigilante-geotecnico"
Issues = "https://github.com/tu-usuario/vigilante-geotecnico/issues"

[project.scripts]
vigilante-geotecnico = "agente_geotecnico:main"
agno-team = "agno_team:agent_os"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"web" = ["*.html", "*.css", "*.js"]

[tool.black]
line-length = 120
exclude = '''/(
        venv
    | \.venv
    | env
    | build
    | dist
    | \.ipynb_checkpoints
    | t_api\.py
    )/'''

[tool.isort]
profile = "black"
line_length = 120
skip = ["t_api.py"]

[tool.ruff]
line-length = 120
extend-exclude = ["venv", ".venv", "env", "build", "dist", ".ipynb_checkpoints", "*.ipynb", "t_api.py"]
